Refactor handling of discount plans
- Generalize find-or-create Plan logic. Have a dedicated plan for our
  GitHub Students deal.
- Move more code to central Marketplace module
- Rename Sync module
