+++ app/main.hs
@@ -1,4 +1,7 @@
-module Main (main) where
+module Main
+    ( main
+    )
+where
 
 import Restyled.Prelude
 
@@ -6,7 +9,7 @@ import LoadEnv (loadEnvFrom)
 import Restyled.Application (runWaiApp)
 import Restyled.Backend.Application
 import Restyled.Backend.Foundation (loadBackend, loadBackendHandle)
-import Restyled.Backend.Marketplace (runSynchronize)
+import Restyled.Backend.MarketplaceSync (runSynchronize)
 import Restyled.Backend.Reconcile (safelyReconcile)
 import Restyled.Development.Seeds (seedDB)
 import Restyled.Export (runExport)
