+++ src/Restyled/GitHubStudents.hs
@@ -2,6 +2,7 @@
 
 module Restyled.GitHubStudents
     ( giftGitHubStudents
+    , githubStudentsPlan
     )
 where
 
@@ -12,13 +13,16 @@ import qualified Data.Text as T
 import Network.HTTP.Simple
 import Network.HTTP.Types.Header (hAccept, hAuthorization)
 import qualified Network.OAuth.OAuth2 as OAuth2
-import Restyled.Backend.DiscountMarketplacePlan
+import Restyled.Marketplace
+import Restyled.Models
+import Restyled.PrivateRepoAllowance
 import Restyled.Yesod
 import qualified Yesod.Auth.OAuth2.GitHubStudents as GitHubStudents
 
 data GitHubStudent = GitHubStudent
     { id :: GitHubUserId
     , login :: GitHubUserName
+    , email :: Maybe Text
     }
     deriving stock Generic
     deriving anyclass FromJSON
@@ -45,9 +49,28 @@ displayGitHubStudent GitHubStudent { id, login } verified = T.unwords
     ]
 
 handleGitHubStudent :: MonadIO m => GitHubStudent -> Bool -> SqlPersistT m ()
-handleGitHubStudent GitHubStudent { id, login } verified
-    | verified = giftDiscountMarketplacePlan id login
-    | otherwise = ungiftDiscountMarketplacePlan login
+handleGitHubStudent GitHubStudent { id, login, email } verified =
+    findOrCreateMarketplacePlan githubStudentsPlan >>= go
+  where
+    go planId
+        | verified = addAccount planId
+        | otherwise = removeAccount planId
+
+    addAccount (Entity planId _) = void $ upsert
+        MarketplaceAccount
+            { marketplaceAccountGithubId = Just id
+            , marketplaceAccountGithubLogin = login
+            , marketplaceAccountMarketplacePlan = planId
+            , marketplaceAccountGithubType = "User"
+            , marketplaceAccountEmail = email
+            , marketplaceAccountBillingEmail = email
+            }
+        [MarketplaceAccountMarketplacePlan =. planId]
+
+    removeAccount (Entity planId _) = deleteWhere
+        [ MarketplaceAccountGithubLogin ==. login
+        , MarketplaceAccountMarketplacePlan ==. planId
+        ]
 
 verifyIsGitHubStudent
     :: (MonadIO m, MonadLogger m) => OAuth2.AccessToken -> m Bool
@@ -65,3 +88,11 @@ verifyIsGitHubStudent token = do
             $ addRequestHeader hAuthorization auth
             $ parseRequest_ "https://education.github.com/api/user"
     auth = "token " <> encodeUtf8 (OAuth2.atoken token)
+
+githubStudentsPlan :: MarketplacePlan
+githubStudentsPlan = MarketplacePlan
+    { marketplacePlanGithubId = Nothing
+    , marketplacePlanPrivateRepoAllowance = PrivateRepoAllowanceUnlimited
+    , marketplacePlanName = "GitHub Students"
+    , marketplacePlanDescription = "Free Unlimited for verified GitHub Students"
+    }
